requirements.txt
django==4.0.3
django-image-cropping==1.7